Scrapped the broken collectionMenu method and created a new one that works. Next, a way to dynamically create new collection items will be added.
Also modified accountMenu by changing variable names for consistency and clarity, as well as changed the positioning of the popup menu to fit to the left side of the account button. This fixes the issue where the popup would appear outside the frame when not in fullscreen.

Finally, the framework for a new filterMenu method was created (mostly a copy of the collectionMenu but with checkboxes).