Added a "Create a New Collection" button within the collection menu and connected it to a new "Create a New Collection" JFrame where users can input a string. Also created a JButton (within the new JFrame) and ActionListener to read in the string from the JTextField for when we connect to the collection model.
Added a question mark onto the MovieDisplay for when a movie does not have an image to display. This more clearly shows that an image is not available than the simple grey rectangle that existed before.

Resized the movieGrid display to better fit the fullscreen view.
Added a dynamically resizing movieGrid display, but commented it out within MovieDisplay.java because of bugs and the fact that we probably won't want the movie images to scale that way. This is something we'll need to come back to later.

Changed the "header" JPanel to a JMenuBar for simplicity and clarity. Looks and functions the exact same though.

Other small changes to existing code for simplification.